+++ src/main/java/brave/btc/dto/app/auth/sms/SmsRequestDto.java
@@ -1,4 +1,4 @@
-package brave.btc.dto.app.auth.register;
+package brave.btc.dto.app.auth.sms;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Pattern;
