+++ src/main/java/brave/btc/exception/sms/SmsCertificationNumberExpiredException.java
@@ -1,4 +1,4 @@
-package brave.btc.exception.auth;
+package brave.btc.exception.sms;
 
 public class SmsCertificationNumberExpiredException extends RuntimeException{
     public SmsCertificationNumberExpiredException(String message) {
