+++ src/main/java/brave/btc/exception/sms/SmsCertificationNumberNotSameException.java
@@ -1,4 +1,4 @@
-package brave.btc.exception.auth;
+package brave.btc.exception.sms;
 
 public class SmsCertificationNumberNotSameException extends RuntimeException{
     public SmsCertificationNumberNotSameException(String message) {
