+++ src/main/java/brave/btc/service/app/auth/SmsService.java
@@ -1,0 +1,36 @@
+package brave.btc.service.app.auth;
+
+import brave.btc.dto.CommonResponseDto;
+
+import java.sql.Timestamp;
+
+public interface SmsService {
+
+    /**
+     * 인증번호를 전송한다.
+     * @param phoneNumber 인증할 전화번호
+     * @return 응답 메세지
+     */
+    CommonResponseDto<Object> sendAuthNumber(String phoneNumber);
+
+    /**
+     * 인증번호를 저장한다.
+     * @param phoneNumber 저장할 전화번호
+     * @param authNumber 저장할 인증번호
+     */
+    void saveCertificationNumber(String phoneNumber, String authNumber);
+
+    /**
+     * 인증번호를 확인한다.
+     * @param phoneNumber 전화번호로 서버애 저장된 인증번호를 불러온다
+     * @param authNumber 요청받은 인증번호
+     */
+    CommonResponseDto<Object> checkAuthNumber(String authNumber, String phoneNumber);
+
+    /**
+     * 인증번호 만료시간을 확인한다.
+     * @param timestamp 인증을 요청한 시각
+     * @param validationTime 인증 유효 시간
+     */
+    void checkCertificationTime(Timestamp timestamp, int validationTime);
+}
